Discard unmatched packets with seqNum == 0
On Linux some percentage of the tranfers passed to the Unpacker will
include packets with seqNum == 0. If a sequence number over 2^15  and 0
are passed to lessThanHandleOverflow() then its result is inverted which
results in the incorrect packet getting discarded corrupting the
synchronization.

At ~1hz samping rate the sequence numbers exceed 2^15 around 30 sec.
The helloSoundplane test duration has been increased to 60 sec according.